+++ R/create_filter_table.R
@@ -71,7 +71,7 @@ create_filter_table <- function(filters,date_filter=FALSE,dsd=NULL,exact_match=T
   if (!date_filter & is.null(dsd)){
     message("The DSD is missing from the create_filter_table function.")
   } else{
-    if (verbose) {message("create_filter_table - filters class: ",class(filters),"; size: ",length(filters),"; filters:",filters)}
+    if (verbose) {message("\ncreate_filter_table - filters class: ",class(filters),"; size: ",length(filters),"; filters:",filters)}
     if (class(filters)=="name") {
       try(filters<-local(filters),silent=verbose)
     }